Reestruturando main
Ajuste de disposição de arquivos
