+++ app/database.py
@@ -6,5 +6,11 @@ DATABASE_URL = "postgresql://postgres:postgres@localhost/datamed"
 
 engine = create_engine(DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
 Base = declarative_base()
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
